backend/requirements.txt
django
django-stubs
djangorestframework
djangorestframework-stubs
mypy
flake8
django-cors-headers
python-dotenv
psycopg2